add logout controller
In order to remove a jwt token this api was created
